split my_page module off of grids
