+++ Gemfile.modules
@@ -40,6 +40,7 @@ group :opf_plugins do
 	gem 'openproject-ldap_groups',               path: 'modules/ldap_groups'
 
 	gem 'grids',                                 path: 'modules/grids'
+	gem 'my_page',                               path: 'modules/my_page'
 	gem 'openproject-boards',                    path: 'modules/boards'
 
 	gem 'openproject-bim_seeder',                path: 'modules/bim_seeder', require: !!(ENV['OPENPROJECT_EDITION'] == 'bim')
