+++ modules/grids/spec/contracts/grids/create_contract_spec.rb
@@ -38,9 +38,11 @@ describe Grids::CreateContract do
   it_behaves_like 'shared grid contract attributes'
 
   describe 'type' do
+    let(:grid) { FactoryBot.build_stubbed(:grid, default_values) }
+
     it_behaves_like 'is writable' do
       let(:attribute) { :type }
-      let(:value) { 'Grids::MyPage' }
+      let(:value) { 'Grids::Grid' }
     end
   end
 
@@ -51,15 +53,6 @@ describe Grids::CreateContract do
       let(:attribute) { :user_id }
       let(:value) { 5 }
     end
-
-    context 'for a Grids::MyPage' do
-      let(:grid) { FactoryBot.build_stubbed(:my_page, default_values) }
-
-      it_behaves_like 'is writable' do
-        let(:attribute) { :user_id }
-        let(:value) { 5 }
-      end
-    end
   end
 
   describe 'project_id' do
@@ -68,15 +61,6 @@ describe Grids::CreateContract do
     it_behaves_like 'is not writable' do
       let(:attribute) { :project_id }
       let(:value) { 5 }
-    end
-
-    context 'for a Grids::MyPage' do
-      let(:grid) { FactoryBot.build_stubbed(:my_page, default_values) }
-
-      it_behaves_like 'is not writable' do
-        let(:attribute) { :project_id }
-        let(:value) { 5 }
-      end
     end
   end
 
@@ -104,12 +88,12 @@ describe Grids::CreateContract do
 
         allow(Grids::Configuration)
           .to receive(:allowed_widget?)
-          .with(Grids::MyPage, :widget1, user)
+          .with(Grids::Grid, :widget1, user)
           .and_return(true)
 
         allow(Grids::Configuration)
           .to receive(:allowed_widget?)
-          .with(Grids::MyPage, :widget2, user)
+          .with(Grids::Grid, :widget2, user)
           .and_return(false)
 
         expect(instance.assignable_values(:widgets, user))
