+++ modules/grids/spec/lib/api/v3/grids/grid_payload_representer_parsing_spec.rb
@@ -71,7 +71,7 @@ describe ::API::V3::Grids::GridPayloadRepresenter, 'parsing' do
       ],
       "_links" => {
         "scope" => {
-          "href" => my_page_path
+          "href" => 'some_path'
         }
       }
     }
@@ -82,7 +82,7 @@ describe ::API::V3::Grids::GridPayloadRepresenter, 'parsing' do
       it 'updates page' do
         grid = representer.from_hash(hash)
         expect(grid.scope)
-          .to eql(my_page_path)
+          .to eql('some_path')
       end
     end
   end
