+++ modules/grids/spec/requests/api/v3/grids/grids_create_form_resource_spec.rb
@@ -61,16 +61,6 @@ describe "POST /api/v3/grids/form", type: :request, content_type: :json do
         .at_path('_type')
     end
 
-    it 'contains a Schema embedding the available values' do
-      expect(subject.body)
-        .to be_json_eql("Schema".to_json)
-        .at_path('_embedded/schema/_type')
-
-      expect(subject.body)
-        .to be_json_eql(my_page_path.to_json)
-        .at_path('_embedded/schema/scope/_links/allowedValues/0/href')
-    end
-
     it 'contains default data in the payload' do
       expected = {
         "rowCount": 4,
@@ -94,136 +84,6 @@ describe "POST /api/v3/grids/form", type: :request, content_type: :json do
     it 'does not have a commit link' do
       expect(subject.body)
         .not_to have_json_path('_links/commit')
-    end
-
-    context 'with /my/page for the scope value' do
-      let(:params) do
-        {
-          '_links': {
-            'scope': {
-              'href': my_page_path
-            }
-          }
-        }
-      end
-
-      it 'contains default data in the payload' do
-        expected = {
-          "rowCount": 7,
-          "columnCount": 4,
-          "options": {},
-          "widgets": [
-            {
-              "_type": "GridWidget",
-              identifier: 'work_packages_assigned',
-              "options": {},
-              startRow: 1,
-              endRow: 7,
-              startColumn: 1,
-              endColumn: 3
-            },
-            {
-              "_type": "GridWidget",
-              identifier: 'work_packages_created',
-              "options": {},
-              startRow: 1,
-              endRow: 7,
-              startColumn: 3,
-              endColumn: 5
-            }
-          ],
-          "_links": {
-            "scope": {
-              "href": "/my/page",
-              "type": "text/html"
-            }
-          }
-        }
-
-        expect(subject.body)
-          .to be_json_eql(expected.to_json)
-          .at_path('_embedded/payload')
-      end
-
-      it 'has no validationErrors' do
-        expect(subject.body)
-          .to be_json_eql({}.to_json)
-          .at_path('_embedded/validationErrors')
-      end
-
-      it 'has a commit link' do
-        expect(subject.body)
-          .to be_json_eql(api_v3_paths.grids.to_json)
-          .at_path('_links/commit/href')
-      end
-    end
-
-    context 'with an unsupported widget identifier' do
-      let(:params) do
-        {
-          '_links': {
-            'scope': {
-              'href': my_page_path
-            }
-          },
-          "widgets": [
-            {
-              "_type": "GridWidget",
-              "identifier": "bogus_identifier",
-              "startRow": 4,
-              "endRow": 5,
-              "startColumn": 1,
-              "endColumn": 2
-            }
-          ]
-        }
-      end
-
-      it 'has a validationError on widget' do
-        expect(subject.body)
-          .to be_json_eql("Widgets is not set to one of the allowed values.".to_json)
-          .at_path('_embedded/validationErrors/widgets/message')
-      end
-    end
-
-    context 'with name set' do
-      let(:params) do
-        {
-          name: 'My custom grid 1',
-          '_links': {
-            'scope': {
-              'href': my_page_path
-            }
-          }
-        }
-      end
-
-      it 'feeds it back' do
-        expect(subject.body)
-          .to be_json_eql("My custom grid 1".to_json)
-          .at_path('_embedded/payload/name')
-      end
-    end
-
-    context 'with options set' do
-      let(:params) do
-        {
-          options: {
-            foo: 'bar'
-          },
-          '_links': {
-            'scope': {
-              'href': my_page_path
-            }
-          }
-        }
-      end
-
-      it 'feeds them back' do
-        expect(subject.body)
-          .to be_json_eql("bar".to_json)
-          .at_path('_embedded/payload/options/foo')
-      end
     end
   end
 end
