+++ modules/grids/spec/services/grids/set_attributes_service_spec.rb
@@ -56,7 +56,7 @@ describe Grids::SetAttributesService, type: :model do
                         contract_class: contract_class)
   end
   let(:call_attributes) { {} }
-  let(:grid_class) { Grids::MyPage }
+  let(:grid_class) { Grids::Grid }
   let(:grid) do
     FactoryBot.build_stubbed(grid_class.name.demodulize.underscore.to_sym, widgets: [])
   end
