+++ modules/my_page/lib/my_page.rb
@@ -1,0 +1,4 @@
+require "my_page/engine"
+
+module MyPage
+end
