+++ modules/my_page/lib/my_page/engine.rb
@@ -1,0 +1,11 @@
+module MyPage
+  class Engine < ::Rails::Engine
+    isolate_namespace MyPage
+
+    include OpenProject::Plugins::ActsAsOpEngine
+
+    config.to_prepare do
+      MyPage::GridRegistration.register!
+    end
+  end
+end
