+++ modules/my_page/spec/models/grids/my_page_spec.rb
