+++ modules/my_page/spec/queries/grids/filters/scope_filter_spec.rb
